ModelServer: add health-check
