+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Note that since we don't clearly distinguish between a public and private interf
 - Add Zhang-Skolnick secondary-structure assignment method which handles coarse-grained models (#49)
 - Calculate bonds for coarse-grained models
 - VolumeServer: Add `health-check` endpoint + `healthCheckPath` config prop to report service health
+- ModelServer: Add `health-check` endpoint + `healthCheckPath` config prop to report service health
 
 ## [v4.5.0] - 2024-07-28
