+++ src/servers/volume/server/web-api.ts
@@ -5,11 +5,9 @@
  *
  * @author David Sehnal <david.sehnal@gmail.com>
  * @author Alexander Rose <alexander.rose@weirdbyte.de>
- * @author Sebastian Bittrich <sebastian.bittrich@rcsb.org>
  */
 
 import * as express from 'express';
-import { promises, constants } from 'fs';
 
 import * as Api from './api';
 import * as Data from './query/data-model';
@@ -20,6 +18,7 @@ import { LimitsConfig, ServerConfig } from '../config';
 import { interpolate } from '../../../mol-util/string';
 import { getSchema, shortcutIconLink } from './web-schema';
 import { swaggerUiIndexHandler, swaggerUiAssetsHandler } from '../../common/swagger-ui';
+import { healthCheck } from '../../common/util';
 
 export function init(app: express.Express) {
     app.locals.mapFile = getMapFileFn();
@@ -34,8 +33,8 @@ export function init(app: express.Express) {
     // Cell /:src/:id/cell/?text=0|1&space=cartesian|fractional
     app.get(makePath(':source/:id/cell/?'), (req, res) => queryBox(req, res, getQueryParams(req, true)));
 
-    // Reports server health depending on `healthCheckPaths` config prop
-    app.get(makePath('health-check'), (_, res) => healthCheck(res));
+    // Reports server health depending on `healthCheckPath` config prop
+    app.get(makePath('health-check'), (_, res) => healthCheck(res, ServerConfig.healthCheckPath));
 
     app.get(makePath('openapi.json'), (req, res) => {
         res.writeHead(200, {
@@ -211,33 +210,4 @@ function queryDone() {
     if (State.shutdownOnZeroPending) {
         process.exit(0);
     }
-}
-
-async function healthCheck(res: express.Response) {
-    if (ServerConfig.healthCheckPath?.length === 0) {
-        healthCheckResponse(res, true);
-        return;
-    }
-
-    for (const path of ServerConfig.healthCheckPath) {
-        try {
-            // assert readable file
-            await promises.access(path, constants.R_OK);
-        } catch (e) {
-            ConsoleLogger.error(`Error accessing path ${path}:`, e);
-            healthCheckResponse(res, false, 'Failed to access data from file system.');
-            return;
-        }
-    }
-    healthCheckResponse(res, true);
-}
-
-function healthCheckResponse(res: express.Response, success: boolean, msg?: string) {
-    res.writeHead(success ? 200 : 500, {
-        'Content-Type': 'text/plain',
-        'Access-Control-Allow-Origin': '*',
-        'Access-Control-Allow-Headers': 'X-Requested-With',
-    });
-    res.write(success ? msg || 'true' : msg || 'false');
-    res.end();
 }
